// src/interface.ts
export interface IReview {
  id: number;
  title: string;
  text: string;
  userNickname: string;
  formattedDate: string;
  rating: number;
  badges: string[];
  positiveFeedbackCount: number;
  negativeFeedbackCount: number;
}

export enum EnumHelpful {
  Default = 0,
  Up = 1,
  Down = 2,
}

export enum EnumFilter {
  Color = '138698751%2C142494633%2C143014975%2C143602960%2C141772550%2C142130291%2C138839801%2C144110492%2C139093884%2C138625697%2C141743948%2C138724049%2C140223907%2C140337792%2C139151262%2C138063095%2C138682627%2C142903080',
  Satisfaction = '138417354%2C142494633%2C138796046%2C138625697%2C143602960%2C144110492%2C138745694%2C143014975%2C138629570%2C138949643%2C144097041%2C139151262%2C138633590%2C142903080%2C141383922%2C138611514%2C142747246',
  Design = '142494633%2C142845232%2C143026389%2C140223907%2C140399132%2C141772550%2C138812542%2C139093884%2C142935560%2C143211023%2C138698751',
  Comfort = '142747246%2C143026389%2C140900593%2C139093884%2C140399132%2C139012144%2C138392963%2C140568968%2C143070081%2C138698751%2C138354714%2C138625697%2C142130291%2C143503895%2C141743948%2C142333741%2C143211023%2C138834951%2C138745694%2C141772550%2C144163849%2C143638983%2C142845232%2C138611514%2C142843369',
  Quality = '138620970%2C138834951%2C138625697%2C142584613%2C140568968%2C143026389%2C141107483'
}

export enum EnumSort {
  Relevant = 'Relevant',
  Newest = 'Newest',
  Helpful = 'Helpful',
  Highest = 'Highest rated',
  Lowest = 'Lowest rated'
}